// src/generated/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const StoryStatus = {
  DRAFT: 'DRAFT',
  PUBLISHED: 'PUBLISHED',
  DELETED: 'DELETED',
  PRIVATE: 'PRIVATE',
  SCHEDULED: 'SCHEDULED'
} as const

export type StoryStatus = (typeof StoryStatus)[keyof typeof StoryStatus]


export const PurchaseMedium = {
  STRIPE: 'STRIPE',
  ESEWA: 'ESEWA',
  KHALTI: 'KHALTI',
  IMEPAY: 'IMEPAY'
} as const

export type PurchaseMedium = (typeof PurchaseMedium)[keyof typeof PurchaseMedium]


export const TransactionType = {
  PURCHASE: 'PURCHASE',
  CHAPTER_UNLOCK: 'CHAPTER_UNLOCK',
  MONTHLY_BONUS: 'MONTHLY_BONUS',
  REFERRAL_BONUS: 'REFERRAL_BONUS',
  SUBSCRIPTION: 'SUBSCRIPTION',
  SUBSCRIPTION_ENDED: 'SUBSCRIPTION_ENDED'
} as const

export type TransactionType = (typeof TransactionType)[keyof typeof TransactionType]


export const AccountStatus = {
  ACTIVE: 'ACTIVE',
  BANNED: 'BANNED',
  DELETED: 'DELETED'
} as const

export type AccountStatus = (typeof AccountStatus)[keyof typeof AccountStatus]


export const Language = {
  English: 'English',
  Spanish: 'Spanish',
  French: 'French',
  German: 'German',
  Italian: 'Italian',
  Portuguese: 'Portuguese',
  Russian: 'Russian',
  Chinese: 'Chinese',
  Japanese: 'Japanese',
  Korean: 'Korean'
} as const

export type Language = (typeof Language)[keyof typeof Language]


export const ChapterPricePool = {
  POOL_50: 'POOL_50',
  POOL_70: 'POOL_70',
  POOL_110: 'POOL_110',
  POOL_150: 'POOL_150'
} as const

export type ChapterPricePool = (typeof ChapterPricePool)[keyof typeof ChapterPricePool]


export const TrafficSource = {
  DIRECT: 'DIRECT',
  SEARCH: 'SEARCH',
  REFERRAL: 'REFERRAL',
  NOTIFICATION: 'NOTIFICATION',
  ADVERTISEMENT: 'ADVERTISEMENT',
  FEED: 'FEED'
} as const

export type TrafficSource = (typeof TrafficSource)[keyof typeof TrafficSource]